Grouped all states by super state
