+++ catkin_ws/src/acr_core/src/core/main.cpp
@@ -2,7 +2,7 @@
 #include "sensorData.cpp"
 #include "states/state.hpp"
 
-#include "states/state_idle.cpp"
+#include "states/idle/state_idle.cpp"
 
 using namespace ros;
