+++ catkin_ws/src/acr_core/src/core/states/idle/state_idle.hpp
@@ -3,7 +3,7 @@
 
 class State_Navigating;
 
-#include "../sensorData.hpp"
+#include "../../sensorData.hpp"
 #include "ros/ros.h"
 
 /* Class representing the idle state.
