+++ catkin_ws/src/acr_core/src/core/states/interacting/state_interact.hpp
@@ -1,26 +1,25 @@
-#ifndef State_Navigating_H
-#define State_Navigating_H
+#ifndef State_Interact_H
+#define State_Interact_H
 
-#include "../sensorData.hpp"
+#include "../../sensorData.hpp"
 class State_Idle;
 
-#include "../../consts.hpp"
 #include "ros/ros.h"
 
-
-/* Abstract superclass for Navigating behaviour.
+/* Abstract superclass for Interact behaviour.
  * This is useful you want certain condition to hold for all States 
  * within the Interact superbehaviour
  * For example: when you want all subbehaviours to return to the Idle state
  * under given conditions
  */ 
-class State_Navigating : public State {	
-
+class State_Interact : public State {	
+ public:
+ 
 	/* Condition under which to go back to the Idle state
 	 */
-	public:
-		bool backToIdle() {return false;};
+	bool backToIdle() {
+		return false;
+	};
 	
 };
-
 #endif
