+++ catkin_ws/src/acr_core/src/core/states/navigating/state_approaching.cpp
@@ -1,6 +1,6 @@
 #include "state_approaching.hpp"
 
-#include "state_concreteInteraction.cpp"
+#include "../interacting/state_concreteInteraction.cpp"
 
 State_Approaching::State_Approaching() { }
